source/basic/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
azure-core==1.30.2
azure-identity==1.17.1
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
distro==1.9.0
gitdb==4.0.11
GitPython==3.1.41
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
msal==1.29.0
msal-extensions==1.2.0
openai==1.35.4
portalocker==2.10.0
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
PyJWT==2.8.0
requests==2.32.3
setuptools==69.0.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.2
wheel==0.43.0
